fix: move factory session fixtures into fixtures subpackage
Relocate published catalog helpers and direct tests under
pkg/factorysessionexecution/fixtures to satisfy pkg-file-count and
backend-size gates while keeping behavioral coverage intact.
